abstract out add char and create a few basic implementions to handle different scenarios and integrate with the framework
abstract out add char and create a few basic implementions to handle different scenarios and integrate with the framework.
compleatly clean up and gut the functionality compaired to the origional all doing  addChar implementation.
there needs to be an array ofd different handlers instead of creating and destroying them all the time and textregion needs to become a stack inatead of being createwd and destroyed all the time. this should give us a good performance boost and allow other matreix to be calculated based upon the l,ayout of the whole document.